\name{@ANNOBJPREFIX@.db}
\alias{@ANNOBJPREFIX@.db}
\alias{@ANNOBJPREFIX@}
\title{Bioconductor annotation data package}
\description{

  Welcome to the @ANNOBJPREFIX@.db annotation Package.  This is an
  platform specific package (sometimes called a chip or ChipDb based
  package).  The purpose is to provide detailed information about the
  platform by mapping probes to genes.  The platform is usually
  abbreviated by the package name @ANNOBJPREFIX@.
  
  To learn more about this package, users are encouraged to learn about
  the select, columns, keys and keytypes methods.  These are described
  in a walkthrough on the bioconductor website as well as in the manual
  pages and vignettes in the AnnotationDbi package.
}
\keyword{datasets}

